config: separate main page components
