+++ src/app/(main)/_components/RightSection/PointRankingItem.tsx
