+++ src/app/(main)/_components/RightSection/RightSection.tsx
@@ -1,0 +1,11 @@
+import PointRankingContainer from "@/app/(main)/_components/RightSection/PointRankingContainer";
+
+function RightSection() {
+  return (
+    <section className="flex flex-col">
+      <PointRankingContainer />
+    </section>
+  );
+}
+
+export default RightSection;
