+++ src/app/(main)/_components/TopSection/TopSection.tsx
@@ -1,0 +1,5 @@
+function TopSection() {
+  return <></>;
+}
+
+export default TopSection;
